Use pytest.mark.freeze_time in plaato tests
